Move hard-disk logic to Storage
