+++ data/userData.csv
@@ -1,1 +1,2 @@
-Todo,game,false
+Todo,game,true
+Deadline,homework,false,2025-12-12
